+++ src/JWT/UnixTimestampDates.php
@@ -4,12 +4,10 @@ declare(strict_types=1);
 
 namespace T0mmy742\TokenAPI\JWT;
 
-use DateTimeImmutable;
 use Lcobucci\JWT\ClaimsFormatter;
 use Lcobucci\JWT\Token\RegisteredClaims;
 
 use function array_key_exists;
-use function assert;
 
 class UnixTimestampDates implements ClaimsFormatter
 {
@@ -17,7 +15,6 @@ class UnixTimestampDates implements ClaimsFormatter
     {
         foreach (RegisteredClaims::DATE_CLAIMS as $claim) {
             if (array_key_exists($claim, $claims)) {
-                assert($claims[$claim] instanceof DateTimeImmutable);
                 $claims[$claim] = $claims[$claim]->getTimestamp();
             }
         }